chore: Move load functionality out of RepositoryInternals
