+++ packages/core/src/state-management/repository-internals.ts
@@ -5,15 +5,11 @@ import {
   CommitType,
   Context,
   type DiagnosticsLogger,
-  type InternalOpts,
-  type LoadOpts,
   StreamUtils,
-  SyncOptions,
   UnreachableCaseError,
 } from '@ceramicnetwork/common'
 import type { LocalIndexApi } from '@ceramicnetwork/indexing'
 import { ServiceMetrics as Metrics } from '@ceramicnetwork/observability'
-import type { StreamID } from '@ceramicnetwork/streamid'
 import type { CAR } from 'cartonne'
 import type { CID } from 'multiformats/cid'
 import {
@@ -40,12 +36,7 @@ import { StreamUpdater } from '../stream-loading/stream-updater.js'
 import type { AnchorService } from '../anchor/anchor-service.js'
 
 const APPLY_ANCHOR_COMMIT_ATTEMPTS = 3
-const CACHE_HIT_LOCAL = 'cache_hit_local'
-const CACHE_HIT_MEMORY = 'cache_hit_memory'
-const CACHE_HIT_REMOTE = 'cache_hit_remote'
 const ANCHOR_POLL_COUNT = 'anchor_poll_count'
-const DEFAULT_LOAD_OPTS = { sync: SyncOptions.PREFER_CACHE, syncTimeoutSeconds: 3 }
-const STREAM_SYNC = 'stream_sync'
 
 export type RepositoryInternalsParams = {
   anchorRequestStore: AnchorRequestStore
@@ -78,14 +69,6 @@ export class RepositoryInternals {
   #streamLoader: StreamLoader
   #streamUpdater: StreamUpdater
 
-  /**
-   * Keeps track of every pinned StreamID that has had its state 'synced' (i.e. a query was sent to
-   * pubsub requesting the current tip for that stream) since the start of this process. This set
-   * only grows over time, in line with how many pinned streams get queried.
-   * @private
-   */
-  #syncedPinnedStreams: Set<string> = new Set()
-
   #numPendingAnchorSubscriptions = 0
 
   constructor(params: RepositoryInternalsParams) {
@@ -117,131 +100,6 @@ export class RepositoryInternals {
    */
   add(state$: RunningState): void {
     this.#inmemory.set(state$.id.toString(), state$)
-  }
-
-  async fromStreamStateStore(streamId: StreamID): Promise<RunningState | undefined> {
-    const streamState = await this.#pinStore.stateStore.load(streamId)
-    if (streamState) {
-      Metrics.count(CACHE_HIT_LOCAL, 1)
-      const runningState = new RunningState(streamState, true)
-      this.add(runningState)
-      const storedRequest = await this.#anchorRequestStore.load(streamId)
-      if (storedRequest !== null && this.#anchorService) {
-        this.confirmAnchorResponse(runningState, storedRequest.cid)
-      }
-      return runningState
-    } else {
-      return undefined
-    }
-  }
-  /**
-   * Returns a stream from wherever we can get information about it.
-   * Starts by checking if the stream state is present in the in-memory cache, if not then
-   * checks the state store, and finally loads the stream from pubsub.
-   */
-  async load(streamId: StreamID, options: LoadOpts & InternalOpts = {}): Promise<RunningState> {
-    const opts = { ...DEFAULT_LOAD_OPTS, ...options }
-
-    const [state$, synced] = await this.#loadingQ.forStream(streamId).run(async () => {
-      const [existingState$, alreadySynced] = await this._fromMemoryOrStoreWithSyncStatus(streamId)
-
-      switch (opts.sync) {
-        case SyncOptions.PREFER_CACHE:
-        case SyncOptions.SYNC_ON_ERROR: {
-          if (!existingState$) {
-            return [await this._loadStreamFromNetwork(streamId, opts.syncTimeoutSeconds), true]
-          }
-
-          if (alreadySynced) {
-            return [existingState$, alreadySynced]
-          } else {
-            await this._sync(existingState$, opts.syncTimeoutSeconds)
-            return [existingState$, true]
-          }
-        }
-        case SyncOptions.NEVER_SYNC: {
-          if (existingState$) {
-            return [existingState$, alreadySynced]
-          }
-          // TODO(CDB-2761): Throw an error if stream isn't found in cache or state store.
-          return [await this._genesisFromNetwork(streamId), false]
-        }
-        case SyncOptions.SYNC_ALWAYS: {
-          return [
-            await this._resyncStreamFromNetwork(streamId, opts.syncTimeoutSeconds, existingState$),
-            true,
-          ]
-        }
-        default:
-          throw new UnreachableCaseError(opts.sync, 'Invalid sync option')
-      }
-    })
-
-    if (!opts.skipCacaoExpirationChecks) {
-      StreamUtils.checkForCacaoExpiration(state$.state)
-    }
-
-    // TODO(WS1-1269): No need to update state if we loaded from the cache or state store
-    await this._updateStateIfPinned(state$)
-    if (synced && state$.isPinned) {
-      this.markPinnedAndSynced(state$.id)
-    }
-
-    return state$
-  }
-
-  /**
-   * Helper function for loading the state for a stream from either the in-memory cache
-   * or the state store, while also returning information about whether or not the state needs
-   * to be synced.
-   * WARNING: This should only be called from within a thread in the loadingQ!!!
-   *
-   * @param streamId
-   * @returns a tuple whose first element is the state that was loaded, and whose second element
-   *   is a boolean representing whether we believe that state should be the most update-to-date
-   *   state for that stream, or whether it could be behind the current tip and needs to be synced.
-   */
-  async _fromMemoryOrStoreWithSyncStatus(
-    streamId: StreamID
-  ): Promise<[RunningState | null, boolean]> {
-    let stream = this._fromMemory(streamId)
-    if (stream) {
-      return [stream, true]
-    }
-
-    stream = await this.fromStreamStateStore(streamId)
-    if (stream) {
-      return [stream, this.wasPinnedStreamSynced(streamId)]
-    }
-    return [null, false]
-  }
-
-  _fromMemory(streamId: StreamID): RunningState | undefined {
-    const state = this.#inmemory.get(streamId.toString())
-    if (state) {
-      Metrics.count(CACHE_HIT_MEMORY, 1)
-    }
-    return state
-  }
-
-  async _genesisFromNetwork(streamId: StreamID): Promise<RunningState> {
-    const state = await this.#streamLoader.loadGenesisState(streamId)
-    Metrics.count(CACHE_HIT_REMOTE, 1)
-
-    const state$ = new RunningState(state, false)
-    this.add(state$)
-    this.#logger.verbose(`Genesis commit for stream ${streamId.toString()} successfully loaded`)
-    return state$
-  }
-
-  /**
-   * Return a stream, either from cache or re-constructed from state store, but will not load from the network.
-   * Adds the stream to cache.
-   */
-  async fromMemoryOrStore(streamId: StreamID): Promise<RunningState | undefined> {
-    const fromMemory = this._fromMemory(streamId)
-    if (fromMemory) return fromMemory
-    return this.fromStreamStateStore(streamId)
   }
 
   /**
@@ -459,65 +317,6 @@ export class RepositoryInternals {
     }
   }
 
-  /**
-   * Takes a stream state that might not contain the complete log (and might in fact contain only the
-   * genesis commit) and kicks off the process to load and apply the most recent Tip to it.
-   *
-   * @param state$ - Current stream state.
-   * @param syncTimeoutSeconds - How much time do we wait for a response from the network.
-   */
-  async _sync(state$: RunningState, syncTimeoutSeconds: number): Promise<void> {
-    const syncedState = await this.#streamLoader.syncStream(state$.state, syncTimeoutSeconds)
-    state$.next(syncedState)
-    Metrics.count(STREAM_SYNC, 1)
-  }
-
-  /**
-   * Loads a stream that the node has never seen before from the network for the first time.
-   *
-   * @param streamId
-   * @param syncTimeoutSeconds - How much time do we wait for a response from the network.
-   */
-  async _loadStreamFromNetwork(
-    streamId: StreamID,
-    syncTimeoutSeconds: number
-  ): Promise<RunningState> {
-    const state = await this.#streamLoader.loadStream(streamId, syncTimeoutSeconds)
-    Metrics.count(STREAM_SYNC, 1)
-    const newState$ = new RunningState(state, false)
-    this.add(newState$)
-    return newState$
-  }
-
-  /**
-   * When SYNC_ALWAYS is provided, we want to reapply and re-validate
-   * the stream state.  We effectively throw out our locally stored state
-   * as it's possible that the commits that were used to construct that
-   * state are no longer valid (for example if the CACAOs used to author them
-   * have expired since they were first applied to the cached state object).
-   * But if we were the only node on the network that knew about the most recent tip,
-   * we don't want to totally forget about that, so we make sure to apply the previously
-   * known about tip so that it can still be considered alongside whatever tip we learn
-   * about from the network.
-   * @param streamId
-   * @param syncTimeoutSeconds
-   * @param existingState$
-   */
-  async _resyncStreamFromNetwork(
-    streamId: StreamID,
-    syncTimeoutSeconds: number,
-    existingState$: RunningState | null
-  ): Promise<RunningState> {
-    const resyncedState = existingState$
-      ? await this.#streamLoader.resyncStream(streamId, existingState$.tip, syncTimeoutSeconds)
-      : await this.#streamLoader.loadStream(streamId, syncTimeoutSeconds)
-
-    Metrics.count(STREAM_SYNC, 1)
-    const newState$ = new RunningState(resyncedState, false)
-    this.add(newState$)
-    return newState$
-  }
-
   publishTip(state$: RunningState): void {
     this.#dispatcher.publishTip(state$.id, state$.tip, state$.state.metadata.model)
   }
@@ -550,24 +349,5 @@ export class RepositoryInternals {
       await this.#pinStore.add(state$)
     }
     await this.indexStreamIfNeeded(state$)
-  }
-
-  markPinnedAndSynced(streamId: StreamID): void {
-    this.#syncedPinnedStreams.add(streamId.toString())
-  }
-
-  markUnpinned(streamId: StreamID): void {
-    this.#syncedPinnedStreams.delete(streamId.toString())
-  }
-
-  /**
-   * Returns whether the given StreamID corresponds to a pinned stream that has been synced at least
-   * once during the lifetime of this process. As long as it's been synced once, it's guaranteed to
-   * be up to date since we keep streams in the state store up to date when we hear pubsub messages
-   * about updates to them.
-   * @param streamId
-   */
-  wasPinnedStreamSynced(streamId: StreamID): boolean {
-    return this.#syncedPinnedStreams.has(streamId.toString())
   }
 }
